moves createPost to parent and takes out createPost component to allow 2-way data binding without mutating props directly
